customer_360_data_platform/requirements.txt
boto3
confluent-kafka
snowflake-connector-python
fastapi
uvicorn
streamlit
pandas
requests
